Tested with the first test case.

+++ examples/magnetics/ISPMSM/run_nsga2.py
@@ -1,5 +1,8 @@
 import os
+import re
 import shutil
+import subprocess
+import sys
 from pathlib import Path
 from datetime import datetime
 
@@ -16,6 +19,8 @@ from pymoo.operators.repair.rounding import RoundingRepair
 from pymoo.operators.sampling.rnd import IntegerRandomSampling
 from pymoo.termination import get_termination
 from pymoo.optimize import minimize
+from torch.nn.init import constant
+from torchgen.executorch.api.et_cpp import return_names
 
 import calculate_average_torque_and_ripple as objective_function
 import machine_model as model
@@ -33,30 +38,46 @@ if __name__ == '__main__':
         'pole_pairs (int)',  # 5 – convert to int
         'stack_length (int/float)',  # 6 – convert to float
         'winding_scheme (str)',  # 7 – string
-        'shortening (int/float)',  # 8 – convert to float
-        'resolution_angle (int)',  # 9 - convert to int
-        'start_position_angle (int/float)',  # 10 - convert to float
-        'end_position_angle (int/float)',  # 11 - convert to float
-        'resolution_average_ripple (int)',  # 12 – convert to int
-        'start_position_average_ripple (int/float)',  # 13 – convert to float
-        'end_position_average_ripple (int/float)',  # 14 – convert to float
-        'rounding (int)',  # 15 – convert to int
-        'lower_boundaries (list)',  # 16 – convert to list
-        'upper_boundaries (list)',  # 17 – convert to list
-        'population_size (int)',  # 18 – convert to int
-        'number_of_offsprings (int)',  # 19 – convert to int
-        'number_of_generations (int)',  # 20 – convert to int
-        'delete_after (bool)'  # 21 – convert to bool
+        'number_of_coil_turns (int/pos)',  # 8 - convert to int
+        'shortening (int/float)',  # 9 – convert to float
+        'resolution_angle (int)',  # 10 - convert to int
+        'start_position_angle (int/float)',  # 11 - convert to float
+        'end_position_angle (int/float)',  # 12 - convert to float
+        'resolution_average_ripple (int)',  # 13 – convert to int
+        'start_position_average_ripple (int/float)',  # 14 – convert to float
+        'end_position_average_ripple (int/float)',  # 15 – convert to float
+        'rounding (int)',  # 16 – convert to int
+        'lower_boundaries (list)',  # 17 – convert to list
+        'upper_boundaries (list)',  # 18 – convert to list
+        'population_size (int)',  # 19 – convert to int
+        'number_of_offsprings (int)',  # 20 – convert to int
+        'number_of_generations (int)',  # 21 – convert to int
+        'delete_after (bool)',  # 22 – convert to bool
+        'number_of_cores (int/pos)'  # 23 - convert to int
     ]
+
+    def go_back():
+        """Closes the simulation GUI and reopens the selector."""
+        root.destroy()
+        subprocess.run([sys.executable, "run_selector.py"])
 
     # Define paths
     folder_path = Path(__file__).resolve().parent
     results_path = folder_path / 'results'
     results_path.mkdir(exist_ok=True)  # Ensure the 'results' directory exists
 
-    # Generate filename with date
+   # Generate NSGA-II result file path
+    date_str = datetime.today().strftime('%Y%m%d')
+    file_path_res = results_path / f'nsga2_result_{date_str}.csv'
+
+    # Generate a filepath for a .csv that contains all the calculated models.
     date_str = datetime.today().strftime('%Y%m%d')
     file_path_all = results_path / f'nsga2_all_{date_str}.csv'
+
+    if file_path_res.exists():
+        file_path_res.unlink()
+    if file_path_all.exists():
+        file_path_all.unlink()
 
     def process_entries(dict_of_entries):
         """Uses the submitted input to perform the simulation and plot the results."""
@@ -69,50 +90,53 @@ if __name__ == '__main__':
             def __init__(self):
                 super().__init__(n_var=3,
                                  n_obj=2,
-                                 xl=np.array(dict_of_entries[labels[16]]),
-                                 xu=np.array(dict_of_entries[labels[17]]),
+                                 xl=np.array(dict_of_entries[labels[17]]),
+                                 xu=np.array(dict_of_entries[labels[18]]),
                                  vtype=int)
 
             def _evaluate(self, x, out, *args, **kwargs):
                 variables = model.VariableParameters(current=dict_of_entries[labels[0]],
                                                      initial_current_angle=dict_of_entries[labels[1]],
                                                      initial_rotor_position=dict_of_entries[labels[2]],
-                                                     rotor_diameter=dict_of_entries[labels[3]] - x[0] / 10,
+                                                     rotor_diameter=dict_of_entries[labels[3]] + x[0] / 10,
                                                      shaft_diameter=dict_of_entries[labels[4]],
                                                      magnet_width=x[1],
                                                      magnet_height=x[2],
                                                      pole_pairs=dict_of_entries[labels[5]],
                                                      stack_lenght=dict_of_entries[labels[6]],
                                                      winding_scheme=dict_of_entries[labels[7]],
-                                                     shortening=dict_of_entries[labels[8]]
+                                                     number_of_coil_turns=dict_of_entries[labels[8]],
+                                                     shortening=dict_of_entries[labels[9]]
                                                      )
                 print(f'X1: {x[0] / 10}, X2: {x[1]}, X3: {x[2]}')
-                f1, f2, _ = objective_function.average_torque_and_ripple(
+                f1, f2, _, _, _ = objective_function.average_torque_and_ripple(
                     variables,
-                    resolution_angle=dict_of_entries[labels[9]],
-                    start_position_angle=dict_of_entries[labels[10]],
-                    end_position_angle=dict_of_entries[labels[11]],
-                    resolution_average_ripple=dict_of_entries[labels[12]],
-                    start_position_average_ripple=dict_of_entries[labels[13]],
-                    end_position_average_ripple=dict_of_entries[labels[14]],
-                    rounding=dict_of_entries[labels[15]],
-                    delete_after=dict_of_entries[labels[21]],
-                    optimisation=True)
+                    resolution_angle=dict_of_entries[labels[10]],
+                    start_position_angle=dict_of_entries[labels[11]],
+                    end_position_angle=dict_of_entries[labels[12]],
+                    resolution_average_ripple=dict_of_entries[labels[13]],
+                    start_position_average_ripple=dict_of_entries[labels[14]],
+                    end_position_average_ripple=dict_of_entries[labels[15]],
+                    rounding=dict_of_entries[labels[16]],
+                    delete_after=dict_of_entries[labels[22]],
+                    optimisation=True,
+                    cores=dict_of_entries[labels[23]],
+                    file_path=file_path_all)
 
-                out['F'] = [f1, f2]
+                out['F'] = [-f1, f2]
 
         problem = MyProblem()
 
         algorithm = NSGA2(
-            pop_size=dict_of_entries[labels[18]],
-            n_offsprings=dict_of_entries[labels[19]],
+            pop_size=dict_of_entries[labels[19]],
+            n_offsprings=dict_of_entries[labels[20]],
             sampling=IntegerRandomSampling(),
             crossover=SBX(prob=0.9, eta=15, vtype=float, repair=RoundingRepair()),
             mutation=PM(prob=1, eta=20, vtype=float, repair=RoundingRepair()),
             eliminate_duplicates=True
         )
 
-        termination = get_termination("n_gen", dict_of_entries[labels[20]])
+        termination = get_termination("n_gen", dict_of_entries[labels[21]])
 
         res = minimize(problem,
                        algorithm,
@@ -127,30 +151,30 @@ if __name__ == '__main__':
         print('Execution time: ' + str(res.exec_time / 60 / 60) + ' hours')
 
         df = pd.DataFrame({'X1': [i / 10 for i in X[:, 0]], 'X2': X[:, 1], 'X3': X[:, 2],
-                           'AVG': F[:, 0], 'RIP': F[:, 1]})
+                           'AVG': -F[:, 0], 'RIP': F[:, 1]})
 
-        folder_path = Path(__file__).resolve().parent
-        results_path = folder_path / 'results'
-        results_path.mkdir(exist_ok=True)
-        date_str = datetime.today().strftime('%Y%m%d')
         df.to_csv(results_path / f'nsga2_result_{date_str}.csv', index=False)
 
         folder_path = ['ang', 'avg']
 
         for i in folder_path:
-            if os.path.exists(i):
+            if os.path.exists(i) and dict_of_entries[labels[22]]:
                 shutil.rmtree(i)
 
         with open(file_path_all, 'r') as f:
-            df = pd.read_csv(f)
+            df1 = pd.read_csv(f)
+        with open(file_path_res, 'r') as f:
+            df2 = pd.read_csv(f)
 
         # Plot the results
         plt.figure(figsize=(8, 6))
-        plt.scatter(df.iloc[:, -2], df.iloc[:, -1], color='blue')
+        plt.scatter(df1.iloc[:, -2], df1.iloc[:, -1], color='blue', label='dominated')
+        plt.scatter(df2.iloc[:, -2], df2.iloc[:, -1], color='red', label='non-dominated')
         plt.title('Pareto Front')
         plt.xlabel('Average Torque [Nm]')
         plt.ylabel('Torque Ripple [Nm]')
         plt.grid(True)
+        plt.legend()
         plt.show()
 
 
@@ -161,34 +185,115 @@ if __name__ == '__main__':
         """
         # Gather the content from each widget in the entries list.
         input_values = [entry.get() for entry in entries]
-        flow_checker = True
+        valid_inputs = {}
 
+        # Makes the type conversion based on their type hints and handles exceptions.
         try:
             for i, input_value in enumerate(input_values):
                 label_text = labels[i]
-                # Process numeric fields: check for the (int/float) type hint.
-                if "(int/float)" in label_text:
-                    input_values[i] = float(input_value)
+
+                if not input_value.strip():  # Check for empty input
+                    print(f"Error: {label_text} cannot be empty.")
+                    return
+
+                elif "(int/float)" in label_text:
+                    valid_inputs[label_text] = float(input_value)
+
+                elif "(int/float/pos)" in label_text:
+                    value = float(input_value)
+                    if value > 0:
+                        valid_inputs[label_text] = value
+                    else:
+                        print(f"Error: {label_text} must be greater than 0.")
+                        return
+
                 elif "(int)" in label_text:
-                    input_values[i] = int(input_value)
-                elif "(list)" in label_text:
-                    input_values[i] = [int(x) for x in input_value.strip("[]").split(",")]
+                    valid_inputs[label_text] = int(input_value)
+
+                elif "(int/pos)" in label_text:
+                    value = int(input_value)
+                    if value > 0:
+                        valid_inputs[label_text] = value
+                    else:
+                        print(f"Error: {label_text} must be greater than 0.")
+                        return
+
                 elif "(bool)" in label_text:
-                    if input_value == 'True':
-                        input_values[i] = True
-                    elif input_value == 'False':
-                        input_values[i] = False
+                    if input_value.lower() in ["true", "false"]:
+                        valid_inputs[label_text] = input_value.lower() == "true"
+                    else:
+                        print(f"Error: {label_text} must be 'True' or 'False'.")
+                        return
+
+                elif "(list)" in label_text:
+                    if bool(re.fullmatch(r"^(\d+),(\d+),(\d+)$", input_value)):
+                        valid_inputs[label_text] = [float(i) for i in input_value.split(",")]
+                    else:
+                        print(f"Error: {label_text} must be in the format i,j,k where i,j and k are integers.")
+                        return
+
                 else:
-                    input_values[i] = input_value
+                    if bool(re.fullmatch(r"^(?:[A-Ca-c]\|){12}$", input_value)):
+                        valid_inputs[label_text] = input_value
+                    elif bool(re.fullmatch(r"^(?:[A-Ca-c][A-Ca-c]\|){12}$", input_value)):
+                        valid_inputs[label_text] = input_value
+                    elif bool(re.fullmatch(r"^(?:[A-Ca-c]){12}$", input_value)):
+                        valid_inputs[label_text] = input_value
+                    else:
+                        print(f"Error: {label_text} must be a valid winding topology."
+                              f"Check the user guide for more information.")
+                        return
 
-        except ValueError as e:
-            print(f"Error converting {labels[i]}: {e}. Please enter a valid type as specified!")
-            print('--------------------------------------------------------------------------------')
-            flow_checker = False
+        except (ValueError, TypeError) as e:
+            print(f"Error processing {label_text}: {e}. Please enter a valid type as specified!")
+            return
 
-        if flow_checker:
-            dict_of_entries = {key: value for key, value in zip(labels, input_values)}
-            process_entries(dict_of_entries)
+        if valid_inputs[labels[4]] >= valid_inputs[labels[3]]:
+            print(f"Error: {labels[3]} must be larger than {labels[4]}.")
+            return
+
+        elif valid_inputs[labels[3]] >= 45:
+            print(f"Error: {labels[3]} must be lower than 45 millimeters.")
+            return
+
+        elif valid_inputs[labels[11]] >= valid_inputs[labels[12]]:
+            print(f"Error: {labels[12]} must be larger than {labels[11]}.")
+            return
+
+        elif valid_inputs[labels[14]] >= valid_inputs[labels[15]]:
+            print(f"Error: {labels[15]} must be larger than {labels[14]}.")
+            return
+
+        elif valid_inputs[labels[16]] > 10:
+            print(f"Error: {labels[16]} must be lower than 10.")
+            return
+
+        elif any(i > j for i, j in zip(valid_inputs[labels[17]], valid_inputs[labels[18]])):
+            print(f"Error: All values of {labels[17]} must be lower than {labels[18]}.")
+            return
+
+        elif valid_inputs[labels[3]] + valid_inputs[labels[18]][0]/10 >= 45:
+            print(f"Error: {labels[3]} + the first element (X1) of {labels[18]}/10 must be lower than 45 millimeters.")
+            return
+
+        elif valid_inputs[labels[17]][1] <= 0:
+            print(f"Error: the second element (X2) of {labels[17]} must be greater than 0.")
+            return
+
+        elif valid_inputs[labels[18]][1] > (constraint:=360 / (valid_inputs[labels[5]] * 2)):
+            print(f"Error: the second element (X2) of {labels[18]} must be lower than {constraint} degrees.")
+            return
+
+        elif valid_inputs[labels[17]][2] <= 0:
+            print(f"Error:the third element (X3) of {labels[17]} must be greater than 0.")
+            return
+
+        elif valid_inputs[labels[18]][2] > (constraint := (valid_inputs[labels[3]] - valid_inputs[labels[4]]) / 2 - 1.5):
+            print(f"Error: the third element (X2) of {labels[18]} must be lower than {constraint} millimeters.")
+            return
+
+        # If all inputs are valid, proceed with processing
+        process_entries(valid_inputs)
 
 
     # Create the main application window.
@@ -198,7 +303,7 @@ if __name__ == '__main__':
     # List to hold the widget or variable associated with each input.
     # For Entry fields, we store the tk.Entry widget.
     entries = []
-    num_columns = 2  # Two input pairs per row (each pair occupies 2 grid columns).
+    num_columns = 4  # Two input pairs per row (each pair occupies 2 grid columns).
 
     # Loop through each label and create a label widget and its corresponding input widget.
     for i, label_text in enumerate(labels):
@@ -217,6 +322,13 @@ if __name__ == '__main__':
             option_menu.config(width=10)
             option_menu.grid(row=row, column=col_group * 2 + 1, padx=5, pady=5, sticky="w")
             entries.append(var)
+        elif 'number_of_cores (int/pos)' in label_text:
+            var = tk.StringVar(root)
+            var.set("1")  # Default value.
+            option_menu = tk.OptionMenu(root, var, "2", "3", "4", "5", "6", "7", "8")
+            option_menu.config(width=10)
+            option_menu.grid(row=row, column=col_group * 2 + 1, padx=5, pady=5, sticky="w")
+            entries.append(var)
         else:
             entry = tk.Entry(root, width=20)
             entry.grid(row=row, column=col_group * 2 + 1, padx=5, pady=5, sticky="w")
@@ -224,10 +336,14 @@ if __name__ == '__main__':
 
     # Determine the row index for the submit button (just after the last row of inputs).
     submit_row = (len(labels) - 1) // num_columns + 1
+    back_row = (len(labels) - 1) // num_columns + 2
 
     # Create the submit button and position it to span across all input columns.
     submit_button = tk.Button(root, text="RUN SIMULATION", command=submit)
     submit_button.grid(row=submit_row, column=0, columnspan=num_columns * 2, pady=10)
+
+    back_button = tk.Button(root, text="BACK", command=go_back)
+    back_button.grid(row=back_row, column=0, columnspan=num_columns * 2, pady=5)
 
     # Start the GUI event loop.
     root.mainloop()
